+++ CHANGELOG.md
@@ -5,6 +5,15 @@ to [Semantic Versioning](http://semver.org/).
 
 
 
+## Unreleased
+
+### Added
+
+- The `Person` API:
+    - `chance.gender()`
+
+
+
 ## [0.1.0] - 2015-08-14
 
 ### Added